+++ src/stories/InputSection.stories.tsx
@@ -1,0 +1,29 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import { fn } from '@storybook/test';
+import InputSection from '../components/InputSection/InputSection';
+import InputField from '../components/InputField';
+
+const meta = {
+  title: 'InputSection',
+  component: InputSection,
+} satisfies Meta<typeof InputSection>;
+
+export default meta;
+
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  args: {
+    title: 'Input Section Title',
+    description: 'Input Section Description',
+    subtitle: 'Input Section Subtitle',
+    children: (
+      <>
+        <InputField value={1234} onChange={fn()} placeholder="0000" />
+        <InputField value={1234} onChange={fn()} placeholder="0000" />
+        <InputField value={1234} onChange={fn()} placeholder="0000" />
+        <InputField value={1234} onChange={fn()} placeholder="0000" />
+      </>
+    ),
+  },
+};
